python/requirements.txt
opencv-python-headless>=4.9
numpy>=1.24
onnxruntime-gpu>=1.17
